Move affine impls into kernels::affine
